Update project's structure
- Move communication logic into the `handlers` package
- Move useful func into the `utils` package
