+++ .gitignore
@@ -1,1 +1,2 @@
 phones.json
+src/**/__pycache__/
